fix deletion code in MovieController and add non-working draft of code to add movies to watchlist by ID in User and WatchlistController
